refactor: extract field-highlighter core CSS to styles folder
