+++ packages/field-highlighter/src/styles/vaadin-user-tags-overlay-styles.d.ts
@@ -1,12 +1,8 @@
 /**
  * @license
- * Copyright (c) 2021 - 2025 Vaadin Ltd.
+ * Copyright (c) 2000 - 2025 Vaadin Ltd.
  * This program is available under Apache License Version 2.0, available at https://vaadin.com/license/
  */
 import type { CSSResult } from 'lit';
-
-export const fieldOutlineStyles: CSSResult;
-
-export const userTagStyles: CSSResult;
 
 export const userTagsOverlayStyles: CSSResult;
