+++ packages/field-highlighter/src/styles/vaadin-user-tags-overlay-styles.js
@@ -1,0 +1,56 @@
+/**
+ * @license
+ * Copyright (c) 2000 - 2025 Vaadin Ltd.
+ * This program is available under Apache License Version 2.0, available at https://vaadin.com/license/
+ */
+import { css } from 'lit';
+import { overlayStyles } from '@vaadin/overlay/src/vaadin-overlay-styles.js';
+
+export const userTagsOverlay = css`
+  :host {
+    background: transparent;
+    box-shadow: none;
+  }
+
+  :scope [part='overlay'] {
+    box-shadow: none;
+    background: transparent;
+    position: relative;
+    left: -4px;
+    padding: 4px;
+    outline: none;
+    overflow: visible;
+  }
+
+  ::slotted([part='tags']) {
+    display: flex;
+    flex-direction: column;
+    align-items: flex-start;
+  }
+
+  :host([dir='rtl']) [part='overlay'] {
+    left: auto;
+    right: -4px;
+  }
+
+  :scope [part='content'] {
+    padding: 0;
+  }
+
+  :host([opening]),
+  :host([closing]) {
+    animation: 0.14s user-tags-overlay-dummy-animation;
+  }
+
+  @keyframes user-tags-overlay-dummy-animation {
+    0% {
+      opacity: 1;
+    }
+
+    100% {
+      opacity: 1;
+    }
+  }
+`;
+
+export const userTagsOverlayStyles = [overlayStyles, userTagsOverlay];
