+++ packages/field-highlighter/src/vaadin-field-outline.js
@@ -7,10 +7,8 @@ import { html, LitElement } from 'lit';
 import { defineCustomElement } from '@vaadin/component-base/src/define.js';
 import { DirMixin } from '@vaadin/component-base/src/dir-mixin.js';
 import { PolylitMixin } from '@vaadin/component-base/src/polylit-mixin.js';
-import { registerStyles, ThemableMixin } from '@vaadin/vaadin-themable-mixin/vaadin-themable-mixin.js';
-import { fieldOutlineStyles } from './vaadin-field-highlighter-styles.js';
-
-registerStyles('vaadin-field-outline', fieldOutlineStyles, { moduleId: 'vaadin-field-outline-styles' });
+import { ThemableMixin } from '@vaadin/vaadin-themable-mixin/vaadin-themable-mixin.js';
+import { fieldOutlineStyles } from './vaadin-field-outline-styles.js';
 
 /**
  * An element used internally by `<vaadin-field-highlighter>`. Not intended to be used separately.
@@ -23,6 +21,10 @@ registerStyles('vaadin-field-outline', fieldOutlineStyles, { moduleId: 'vaadin-f
 export class FieldOutline extends ThemableMixin(DirMixin(PolylitMixin(LitElement))) {
   static get is() {
     return 'vaadin-field-outline';
+  }
+
+  static get styles() {
+    return fieldOutlineStyles;
   }
 
   static get properties() {
